Replace pythonnet with QuantConnect variant
Differences are hard to identitfy, but QuantConnect was the original
development version. My understanding was that it was forked from
pythonnet and I wanted to use the main project, but it appears
QuantConnect does more implicit conversions I was relying on in testing.
Switching back to QuantConnect version should fix the problems with
certain Clr to Managed runtime exceptions.