Always run Python from the extension directory
The working directory should not be the calling AP Suite working
directory, rather it should temporarily switch the calling extension
directory. This should always be a directory with user permission,
so there should be no issues with writing files to disk if necessary.
It also facilitates finding relative paths to content files that may be
shipped with the extension.