Include assembly qualifier in dialog view ToString
The ToString is used by Prism to resolve the view model
to the view, but this causes issues with different versions
if the class name hasn't changed. Including the versio
qualified assembly should effecively scope each resolution
of each view model to only the view of the same version
and assembly.