Rename views
The new override of view model provider fixed assembly load context
isolation issues going forward, but it continues to conflict with old
versions that don't have the fix. As a one time attempt to mitigate
conflicts between an old version and the new version, also rename
the views so that the sting matching based on view name (which
we do not control as part of the Prism framework) does not override
one requisition from an ALC with another when the view key is the same.