Update utilities to latest version
Version 1.8.0 shipped with Utilities 4.0.0. After publication,
it was realized that this library only depended on 3.7.0, and a quick
hotfix was made to attempt to roll it back. The goal was to try to
maintain maximum compatibility. This ended up causing confusion
as updating from 1.8.0 to 1.8.1 will break builds if Utilities 4.0.0 was
not explicitly included an only inherited from this dependency.
As adoption of this library is not yet widespread, and the expectation
is that the host application should be kept updated when possible, it
is better to simply unlist the hotfix and include the higher Utilities
version despite being unnecessarily high simply to minimize confusion.