Auto-configuration updates
Fixes issues with dialog display in different versions due to global
registration of dialogs instead of assembly load context isolation,
and fix issue with xaml in shared dependency libraries by rewriting
xaml files in pure C# to avoid global BAML loader references in WPF.

Includes a Python not found dialog for configurable distribution mode.
Using the same design as for the legacy Anaconda only mode, so fixed
and generalized the dialog view for use in both the configurable mode
and also in the legacy Anaconda mode, despite the marked deprecation.

Fix issue where configured conda distributions might not be auto-
configured correctly due to using venv path instead of the exe to check
for conda-meta path. The exe should be more reliable.

Updated auto-configuration to do more automatic guidance. If no
distribution is located, display the message to download instead of
an empty list. If only one options, immediately select without even
displaying the selection dialog as there is only one option anyways.

When a distribution is selected, immediately try to run the auto-
configuration instead of having to manually select the button. Assume
more aggressively that we can overwrite existing configurations. This
should be an better UX for most users at the cost of requiring more
manual effort if advanced configurations were manually set, which should
be a less likely case.

Finally, automatically display the venv creation dialog if after auto-
configuration the venv directory does not exists.